build_your_own_IoT_platform/src/apps/generators/requirements.txt
nats-py==2.7.0
pydantic==2.8.2
python-dateutil==2.9.0
cryptography==42.0.7